temperature to double type

+++ lib/screens/weather_screen.dart
@@ -25,7 +25,7 @@ class _WeatherScreenState extends State<WeatherScreen> {
   }
 
   void updateData(dynamic weatherData) {
-    double temperatureDouble = weatherData['main']['temp'];
+    double temperatureDouble = weatherData['main']['temp'].toDouble();
     // temperature = temperatureDouble.toInt(); //toInt() 하는 방법 1 : 소수점 이하를 버림
     temperature = temperatureDouble.round(); //round() 하는 방법 2 : 소수점을 반올림함
     cityName = weatherData['name'];
